feat(terraform-azure): set baseline b2s count for basic cluster.
Magento, by default in the configuration we provision, needs ~4.1 cores and ~16Gb of RAM.

AKS clusters (on B2s) have ~1900 vCPU of and ~2GB of RAM (regardless of what the AKS documentation says).
This difference from documentation is primarily due to the overhead of running K8s nodes.

That said, in reality, after required daemons and deployments are
accounted for, we lose another ~550m of vCPU. This means, in reality,
we only have ~1350vCPU per node ><.

That said, to get within current allocation bounds we need at least 6 nodes.

8 * 1350vCPU = 10800vCPU
8 * 2.1GB = 16.8GB

We're quite overrallocated on vCPU and we're only slightly over-allocated on RAM,
which is less than ideal, but this is the most cost effective
solution that I could come up with ($243.54/mo - PAYG).